# Feature/optimize fetching loading build