Huge PR.
For fetching modification, this enable minimal iterations.

For external JSON and localStorage, after the first grab of files, loading should be a little better (1 to 3 s on my old Win7 computer) 
Teasing : With next PR loading will be much faster (3 to 4 s compared to master) ^^

For weight modification, it is needed to reduce drastically requests to ESI (needed for Nearby region search #145 )